move IllegalStartSkipPolicy to core
