+++ eva-accession-core/src/main/java/uk/ac/ebi/eva/accession/core/batch/policies/IllegalStartSkipPolicy.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.ac.ebi.eva.accession.remapping.batch.policies;
+package uk.ac.ebi.eva.accession.core.batch.policies;
 
 import org.springframework.batch.core.step.skip.SkipLimitExceededException;
 import org.springframework.batch.core.step.skip.SkipPolicy;
