+++ eva-accession-core/src/test/java/uk/ac/ebi/eva/accession/core/batch/policies/IllegalStartSkipPolicyTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.ac.ebi.eva.accession.release.batch.policies;
+package uk.ac.ebi.eva.accession.core.batch.policies;
 
 import org.junit.Before;
 import org.junit.Test;
