--- eva-accession-release/src/main/java/uk/ac/ebi/eva/accession/release/batch/policies/IllegalStartSkipPolicy.java
@@ -1,29 +1,0 @@
-/*
- * Copyright 2019 EMBL - European Bioinformatics Institute
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package uk.ac.ebi.eva.accession.release.batch.policies;
-
-import org.springframework.batch.core.step.skip.SkipLimitExceededException;
-import org.springframework.batch.core.step.skip.SkipPolicy;
-
-import uk.ac.ebi.eva.accession.core.exceptions.PositionOutsideOfContigException;
-
-public class IllegalStartSkipPolicy implements SkipPolicy {
-
-    @Override
-    public boolean shouldSkip(Throwable exception, int skipCount) throws SkipLimitExceededException {
-        return exception instanceof PositionOutsideOfContigException;
-    }
-}
