--- eva-accession-remapping/src/main/java/uk/ac/ebi/eva/accession/remapping/configuration/batch/policies/IllegalStartSkipPolicyConfiguration.java
@@ -1,30 +1,0 @@
-/*
- * Copyright 2019 EMBL - European Bioinformatics Institute
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package uk.ac.ebi.eva.accession.remapping.configuration.batch.policies;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import uk.ac.ebi.eva.accession.remapping.batch.policies.IllegalStartSkipPolicy;
-
-@Configuration
-public class IllegalStartSkipPolicyConfiguration {
-
-    @Bean
-    public IllegalStartSkipPolicy invalidStartSkipPolicy() {
-        return new IllegalStartSkipPolicy();
-    }
-}
