+++ eva-accession-remapping/src/test/java/uk/ac/ebi/eva/accession/remapping/test/configuration/BatchTestConfiguration.java
@@ -24,8 +24,7 @@ import org.springframework.context.annotation.Import;
 import uk.ac.ebi.eva.accession.core.configuration.nonhuman.MongoConfiguration;
 import uk.ac.ebi.eva.accession.remapping.configuration.InputParametersConfiguration;
 import uk.ac.ebi.eva.accession.remapping.configuration.batch.jobs.ExportSubmittedVariantsJobConfiguration;
-import uk.ac.ebi.eva.accession.remapping.configuration.batch.policies.ChunkSizeCompletionPolicyConfiguration;
-import uk.ac.ebi.eva.accession.remapping.configuration.batch.policies.IllegalStartSkipPolicyConfiguration;
+import uk.ac.ebi.eva.accession.remapping.configuration.batch.policies.PoliciesConfiguration;
 import uk.ac.ebi.eva.accession.remapping.runner.AccessionReleaseJobLauncherCommandLineRunner;
 import uk.ac.ebi.eva.commons.batch.job.JobExecutionApplicationListener;
 
@@ -33,8 +32,7 @@ import uk.ac.ebi.eva.commons.batch.job.JobExecutionApplicationListener;
 @Import({MongoConfiguration.class,
          InputParametersConfiguration.class,
          ExportSubmittedVariantsJobConfiguration.class,
-         ChunkSizeCompletionPolicyConfiguration.class,
-         IllegalStartSkipPolicyConfiguration.class,
+         PoliciesConfiguration.class,
          AccessionReleaseJobLauncherCommandLineRunner.class
 })
 public class BatchTestConfiguration {
